Add support for skipping certain kinds of kubernetes configs
Custom resource definitions are at the current time not supported by kubeval. This causes issues for validation of these types and due to this there were already some workarounds for the sealed secret related things.

This PR adds support for defining your own types to skip and removed the old workarounds used for sealed secrets controller and secrets. A benefit with this approach (compared to totally ignoring the files) is that kubeval will still check that the yaml syntax is valid for these files, so it could help catching some syntax errors.