Add support for overriding base level kube files per cluster (#21)
* Add support for overriding base level kube files per cluster

The kube/*.yaml files are all applied to all clusters when running the init_kubernetes task. This makes it possible to per context (cluster) override the files (or add totally new files) to be applied to the cluster. This is done by storing the override files in kube/overrides/<kube_context>/*.yaml

* Update README