Unify order of env/context and overrides
The overrides for components are in paths of the form
`envs/<env>/overrides/...*.yaml`
and the kubernetes overrides were added on the form
`kube/overrides/<kube_context>/*.yaml`.

This fixes it to be consistent, by switching to
`kube/<kube_context>/overrides/*.yaml`, so the `overrides` folder comes after the context/env also here.