Fix seal-secrets --only-changed to work with new secrets files
If you added a new `.unsealed-secrets.yaml` file, the `invoke seal-secrets --only-changed` crashed since it could not open the (non-existing) corresponding sealed version of the file. The right solution in this case is to create the new sealed version of the file and not update any of the other existing unmodified secrets.